Refactor tpm_tis_core.h and move macros and enums to a new header
